[FIX] #14 WeightChangeView 디자인 수정 및 폴더링
